Fix issue with initdb shell function clashing on OS X.
Calls to initdb like "initdb --version" were not working correctly.
The cause was the peg function library including a function named
initdb.  initdb the PostgreSQL binary looks for itself to determine
how to call other programs, and something about that was breaking
down, when the function had the same name.